move fp functions to fp.py
