+++ salmon_lib/parsers/__init__.py
@@ -1,3 +1,4 @@
 from .bse import *
 from .general import *
 from .opt import *
+from .fp import *
